Now that type has changed due to [d0fa9eee56], cast back to match total variable.
FossilOrigin-Name: 77371043a47ed234f78802c8e0a156d24a0f92cf495116f63ede0e71f453e2e1

+++ src/http_socket.c
@@ -203,8 +203,8 @@ size_t socket_send(void *NotUsed, const void *pContent, size_t N){
   while( N>0 ){
     sent = send(iSocket, pContent, N, 0);
     if( sent<=0 ) break;
-    total += sent;
-    N -= sent;
+    total += (size_t)sent;
+    N -= (size_t)sent;
     pContent = (void*)&((char*)pContent)[sent];
   }
   return total;
